// frontend/src/app/modules/admin/interfaces/IGiftProgram.ts
export interface IGiftProgram {
  giftformGuid: string;
  candidateName: string;
  giftformReviewerId: string;
  currentDesignation: string;
  desiredDesignation: string;
  gapsIdentified: string;
  planofAction: string;
  targetAchievementDate: string;
  isSupportRequired: boolean;
  endResult: string;
  majorAchievements: string;
  learningAndTransistionProcess: string;
  careerGrowthContribution: string;
  workRelatedTrainingAndCertifications: string;
  giftformAdminStatus: number;
  giftformManagerStatus: number;
  giftformRejectionReason: string;
}